WIP move saving stuff to new module
