+++ src/server/mod.rs
@@ -16,8 +16,6 @@ extern crate httparse;
 extern crate safemem;
 extern crate twoway;
 
-use tempdir::TempDir;
-
 use std::borrow::Borrow;
 use std::collections::HashMap;
 use std::fs;
@@ -25,12 +23,17 @@ use std::io::prelude::*;
 use std::path::{Path, PathBuf};
 use std::{io, mem};
 
+use tempdir::TempDir;
+
 use self::boundary::BoundaryReader;
 
 use self::field::PrivReadEntry;
 
-pub use self::field::{MultipartField, MultipartFile, MultipartData, ReadEntry, ReadEntryResult,
-                      SaveBuilder, SavedFile};
+pub use self::field::{MultipartField, MultipartFile, MultipartData, ReadEntry, ReadEntryResult};
+
+use self::save::SaveBuilder;
+
+pub use self::save::{Entries, SaveResult};
 
 macro_rules! try_opt (
     ($expr:expr) => (
@@ -73,6 +76,8 @@ pub mod nickel;
 
 #[cfg(feature = "tiny_http")]
 pub mod tiny_http;
+
+pub mod save;
 
 /// The server-side implementation of `multipart/form-data` requests.
 ///
@@ -136,21 +141,18 @@ impl<B: Read> Multipart<B> {
         }
     }
 
+    pub fn save(&mut self) -> SaveBuilder<Self> {
+        SaveBuilder::new(self)
+    }
+
     /// Read the request fully, parsing all fields and saving all files in a new temporary
     /// directory under the OS temporary directory. 
     ///
     /// If there is an error in reading the request, returns the partial result along with the
     /// error. See [`SaveResult`](enum.SaveResult.html) for more information.
+    #[deprecated = "use `.save().temp()` instead"]
     pub fn save_all(&mut self) -> SaveResult {
-        let mut entries = match Entries::new_tempdir() {
-            Ok(entries) => entries,
-            Err(err) => return SaveResult::Error(err),
-        };
- 
-        match self.read_to_entries(&mut entries, None) {
-            Ok(()) => SaveResult::Full(entries),
-            Err(err) => SaveResult::Partial(entries, err),
-        }
+        self.save().temp()
     }
 
     /// Read the request fully, parsing all fields and saving all files in a new temporary
@@ -158,15 +160,11 @@ impl<B: Read> Multipart<B> {
     ///
     /// If there is an error in reading the request, returns the partial result along with the
     /// error. See [`SaveResult`](enum.SaveResult.html) for more information.
+    #[deprecated = "use `.save().with_temp_dir()` instead"]
     pub fn save_all_under<P: AsRef<Path>>(&mut self, dir: P) -> SaveResult {
-        let mut entries = match Entries::new_tempdir_in(dir) {
-            Ok(entries) => entries,
+        match TempDir::new_in(dir, "multipart") {
+            Ok(temp_dir) => self.save().with_temp_dir(temp_dir),
             Err(err) => return SaveResult::Error(err),
-        };
-
-        match self.read_to_entries(&mut entries, None) {
-            Ok(()) => SaveResult::Full(entries),
-            Err(err) => SaveResult::Partial(entries, err),
         }
     }
 
@@ -177,16 +175,9 @@ impl<B: Read> Multipart<B> {
     ///
     /// If there is an error in reading the request, returns the partial result along with the
     /// error. See [`SaveResult`](enum.SaveResult.html) for more information.
+    #[deprecated = "use `.save().limit(limit)` instead"]
     pub fn save_all_limited(&mut self, limit: u64) -> SaveResult {
-        let mut entries = match Entries::new_tempdir() {
-            Ok(entries) => entries,
-            Err(err) => return SaveResult::Error(err),
-        };
-
-        match self.read_to_entries(&mut entries, Some(limit)) {
-            Ok(()) => SaveResult::Full(entries),
-            Err(err) => SaveResult::Partial(entries, err),
-        }
+        self.save().limit(limit).temp()
     }
 
     /// Read the request fully, parsing all fields and saving all files in a new temporary
@@ -196,33 +187,12 @@ impl<B: Read> Multipart<B> {
     ///
     /// If there is an error in reading the request, returns the partial result along with the
     /// error. See [`SaveResult`](enum.SaveResult.html) for more information.
+    #[deprecated = "use `.save().limit(limit).with_temp_dir()` instead"]
     pub fn save_all_under_limited<P: AsRef<Path>>(&mut self, dir: P, limit: u64) -> SaveResult {
-        let mut entries = match Entries::new_tempdir_in(dir) {
-            Ok(entries) => entries,
+        match TempDir::new_in(dir, "multipart") {
+            Ok(temp_dir) => self.save().limit(limit).with_temp_dir(temp_dir),
             Err(err) => return SaveResult::Error(err),
-        };
-
-        match self.read_to_entries(&mut entries, Some(limit)) {
-            Ok(()) => SaveResult::Full(entries),
-            Err(err) => SaveResult::Partial(entries, err),
         }
-    }
-
-    fn read_to_entries(&mut self, entries: &mut Entries, limit: Option<u64>) -> io::Result<()> {
-        while let Some(field) = try!(self.read_entry()) {
-            match field.data {
-                MultipartData::File(mut file) => {
-                    let file = try!(file.save().limit(limit).with_dir(&entries.dir));
-
-                    entries.files.insert(field.name, file);
-                },
-                MultipartData::Text(text) => {
-                    entries.fields.insert(field.name, text.text);
-                },
-            }
-        }
-
-        Ok(())
     }
 }
 
@@ -247,53 +217,6 @@ impl<R: Read> PrivReadEntry for Multipart<R> {
     }
 }
 
-/// The result of [`Multipart::save_all()`](struct.Multipart.html#method.save_all).
-#[derive(Debug)]
-pub enum SaveResult {
-    /// The operation was a total success. Contained are all entries of the request.
-    Full(Entries),
-    /// The operation errored partway through. Contained are the entries gathered thus far,
-    /// as well as the error that ended the process.
-    Partial(Entries, io::Error),
-    /// The `TempDir` for `Entries` could not be constructed. Contained is the error detailing the
-    /// problem.
-    Error(io::Error),
-}
-
-impl SaveResult {
-    /// Take the `Entries` from `self`, if applicable, and discarding
-    /// the error, if any.
-    pub fn to_entries(self) -> Option<Entries> {
-        use self::SaveResult::*;
-
-        match self {
-            Full(entries) | Partial(entries, _) => Some(entries),
-            Error(_) => None,
-        }
-    }
-
-    /// Decompose `self` to `(Option<Entries>, Option<io::Error>)`
-    pub fn to_opt(self) -> (Option<Entries>, Option<io::Error>) {
-        use self::SaveResult::*;
-
-        match self {
-            Full(entries) => (Some(entries), None),
-            Partial(entries, error) => (Some(entries), Some(error)),
-            Error(error) => (None, Some(error)),
-        }
-    }
-
-    /// Map `self` to an `io::Result`, discarding the error in the `Partial` case.
-    pub fn to_result(self) -> io::Result<Entries> {
-        use self::SaveResult::*;
-
-        match self {
-            Full(entries) | Partial(entries, _) => Ok(entries),
-            Error(error) => Err(error),
-        }
-    }
-}
-
 /// A server-side HTTP request that may or may not be multipart.
 ///
 /// May be implemented by mutable references if providing the request or body by-value is
@@ -310,114 +233,4 @@ pub trait HttpRequest {
 
     /// Return the request body for reading.
     fn body(self) -> Self::Body;
-}
-
-/// A result of `Multipart::save_all()`.
-#[derive(Debug)]
-pub struct Entries {
-    /// The text fields of the multipart request, mapped by field name -> value.
-    pub fields: HashMap<String, String>,
-    /// A map of file field names to their contents saved on the filesystem.
-    pub files: HashMap<String, SavedFile>,
-    /// The directory the files in this request were saved under; may be temporary or permanent.
-    pub dir: SaveDir,
-}
-
-impl Entries {
-    fn new_tempdir_in<P: AsRef<Path>>(path: P) -> io::Result<Self> {
-        TempDir::new_in(path, "multipart").map(Self::with_tempdir)
-    }
-
-    fn new_tempdir() -> io::Result<Self> {
-        TempDir::new("multipart").map(Self::with_tempdir)
-    }
-
-    fn with_tempdir(tempdir: TempDir) -> Entries {
-        Entries {
-            fields: HashMap::new(),
-            files: HashMap::new(),
-            dir: SaveDir::Temp(tempdir),
-        }
-    }
-}
-
-/// The save directory for `Entries`. May be temporary (delete-on-drop) or permanent.
-#[derive(Debug)]
-pub enum SaveDir {
-    /// This directory is temporary and will be deleted, along with its contents, when this wrapper
-    /// is dropped.
-    Temp(TempDir),
-    /// This directory is permanent and will be left on the filesystem when this wrapper is dropped.
-    Perm(PathBuf),
-}
-
-impl SaveDir {
-    /// Get the path of this directory, either temporary or permanent.
-    pub fn as_path(&self) -> &Path {
-        use self::SaveDir::*;
-        match *self {
-            Temp(ref tempdir) => tempdir.path(),
-            Perm(ref pathbuf) => &*pathbuf,
-        }
-    }
-
-    /// Returns `true` if this is a temporary directory which will be deleted on-drop.
-    pub fn is_temporary(&self) -> bool {
-        use self::SaveDir::*;
-        match *self {
-            Temp(_) => true,
-            Perm(_) => false,
-        }
-    }
-
-    /// Unwrap the `PathBuf` from `self`; if this is a temporary directory,
-    /// it will be converted to a permanent one.
-    pub fn into_path(self) -> PathBuf {
-        use self::SaveDir::*;
-
-        match self {
-            Temp(tempdir) => tempdir.into_path(),
-            Perm(pathbuf) => pathbuf,
-        }
-    }
-
-    /// If this `SaveDir` is temporary, convert it to permanent.
-    /// This is a no-op if it already is permanent.
-    ///
-    /// ###Warning: Potential Data Loss
-    /// Even though this will prevent deletion on-drop, the temporary folder on most OSes
-    /// (where this directory is created by default) can be automatically cleared by the OS at any
-    /// time, usually on reboot or when free space is low.
-    ///
-    /// It is recommended that you relocate the files from a request which you want to keep to a 
-    /// permanent folder on the filesystem.
-    pub fn keep(&mut self) {
-        use self::SaveDir::*;
-        *self = match mem::replace(self, Perm(PathBuf::new())) {
-            Temp(tempdir) => Perm(tempdir.into_path()),
-            old_self => old_self,
-        };
-    }
-
-    /// Delete this directory and its contents, regardless of its permanence.
-    ///
-    /// ###Warning: Potential Data Loss
-    /// This is very likely irreversible, depending on the OS implementation.
-    ///
-    /// Files deleted programmatically are deleted directly from disk, as compared to most file
-    /// manager applications which use a staging area from which deleted files can be safely
-    /// recovered (i.e. Windows' Recycle Bin, OS X's Trash Can, etc.).
-    pub fn delete(self) -> io::Result<()> {
-        use self::SaveDir::*;
-        match self {
-            Temp(tempdir) => tempdir.close(),
-            Perm(pathbuf) => fs::remove_dir_all(&pathbuf),
-        }
-    }
-}
-
-impl AsRef<Path> for SaveDir {
-    fn as_ref(&self) -> &Path {
-        self.as_path()
-    }
 }
